# service2110/docs/swagger/service2110.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backend API",
    "description": "API документация для Backend проекта на Fastify + TypeScript + PostgreSQL\n\n## Глоссарий терминов\n\n- **ТФР (Территориальный финансовый репозиторий)** - централизованное хранилище финансовых отчётов\n- **DAPP** - Data Application Processing - система обработки данных\n- **FC** - File Conversion - система конвертации файлов",
    "version": "2.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "schemas": {
      "PaginationRequestDto": {
        "type": "object",
        "properties": {
          "page": {
            "default": 1,
            "description": "Номер страницы (начиная с 1)",
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "limit": {
            "default": 20,
            "description": "Размер страницы",
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": [
          "page",
          "limit"
        ],
        "additionalProperties": false
      },
...
      "DateSchema": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Дата в формате YYYY-MM-DD"
      },
...
      "SortOrderEnum": {
        "default": "desc",
        "description": "Направление сортировки",
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "x-enum-descriptions": [
          "Сортировка по возрастанию",
          "Сортировка по убыванию"
        ],
        "x-enum-varnames": [
          "ASC",
          "DESC"
        ],
        "oneOf": [
          {
            "const": "asc",
            "description": "Сортировка по возрастанию",
            "title": "ASC"
          },
          {
            "const": "desc",
            "description": "Сортировка по убыванию",
            "title": "DESC"
          }
        ]
      },
...
      "ШnventoryAccountsListFilterDto": {
        "type": "object",
        "properties": {
          "inventoryOrderId": {
            "description": "ИД приказа инвентаризации",
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "accountSurrogateId": {
            "description": "Суррогатный ключ счёта",
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "accountTypeId": {
            "description": "Тип счёта (одиночный id)",
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "bs2": {
            "description": "БС-2 (массив int)",
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          },
          "accountNum": {
            "description": "Номер счёта (accountNum в DOC)",
            "type": "string"
          },
          "accountType": {
            "description": "Типы счёта (массив)",
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          },
          "responsibleUnit": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "responsibleUnitType": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inventoryAccountStatus": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceBank": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productName": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "manualControlRuleNumber": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          },
          "showZeroBalanceAccounts": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Фильтр для списка счетов"
      },
      "InventorizationGetAccountsListRequestDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "description": "Параметры пагинации",
            "$ref": "#/components/schemas/PaginationRequestDto"
          },
          "sorting": {
            "type": "object",
            "properties": {
              "sortOrder": {
                "$ref": "#/components/schemas/SortOrderEnum"
              },
              "sortBy": {
                "type": "string",
                "enum": [
                  "id",
                  "accountNumber",
                  "accountSurrogateId",
                  "updatedAt"
                ],
                "description": "Колонка для сортировки"
              }
            },
            "required": [
              "sortOrder",
              "sortBy"
            ],
            "additionalProperties": false,
            "description": "Параметры сортировки (колонка — фиксированный набор)"
          },
          "filter": {
            "$ref": "#/components/schemas/ШnventoryAccountsListFilterDto"
          }
        },
        "required": [
          "pagination",
          "sorting",
          "filter"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventorizationAccountDetailDto"
            },
            "description": "Список счетов"
          },
          "totalItems": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Общее количество счетов"
          }
        },
        "required": [
          "items",
          "totalItems"
        ],
        "additionalProperties": false
      },
...
      "InventorizationAccountDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "ИД записи в реестре"
          },
          "accountSurrogateId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Суррогатный ключ счёта (DOC: accountSurrogateKey)"
          },
          "accountNumber": {
            "description": "Номер счёта (DOC: accountNum)",
            "type": "string"
          },
          "accountOpenDate": {
            "$ref": "#/components/schemas/DateSchema"
          },
          "accountType": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "clientName": {
            "type": "string"
          },
          "agreementNum": {
            "type": "string"
          },
          "originalUnit": {
            "type": "string"
          },
          "bs2": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "accountName": {
            "type": "string"
          },
          "lastOperationDate": {
            "$ref": "#/components/schemas/DateSchema"
          },
          "agreementStatus": {
            "type": "string"
          },
          "accountOutRubSum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "accountOutSum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "manualControlRuleNumber": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "manualControlFlag": {
            "type": "boolean"
          },
          "inventoryAccountStatus": {
            "type": "string"
          },
          "reversedDebitSumRub": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "reversedDebitSum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "reversedCreditSumRub": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "reversedCreditSum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "sourceBank": {
            "type": "string"
          },
          "dataDate": {
            "$ref": "#/components/schemas/DateSchema"
          },
          "currencyIsoCode": {
            "type": "string"
          },
          "clientInn": {
            "type": "string"
          },
          "discrepancyDescription": {
            "type": "string"
          },
          "discrepancySum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "discrepancyReason": {
            "type": "string"
          },
          "resolutionActions": {
            "type": "string"
          },
          "resolutionDate": {
            "$ref": "#/components/schemas/DateSchema"
          },
          "manualUnit": {
            "type": "string"
          },
          "updatedAt": {
            "description": "Для сортировки списка",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "accountSurrogateId"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountHistoryItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "accountSurrogateId": {
            "description": "Суррогат счёта (DOC)",
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "changedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "changedBy": {
            "type": "string"
          },
          "fieldName": {
            "type": "string"
          },
          "oldValue": {
            "type": "string"
          },
          "newValue": {
            "type": "string"
          },
          "action": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "changedAt"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountHistoryResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventorizationAccountHistoryItemDto"
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
      "InventorizationManualUnitBulkRequestDto": {
        "type": "object",
        "properties": {
          "accountSurrogateIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          },
          "manualResponsibleUnit": {
            "type": "string"
          },
          "force": {
            "type": "boolean"
          }
        },
        "required": [
          "accountSurrogateIds"
        ],
        "additionalProperties": false
      },
      "InventorizationManualUnitSingleRequestDto": {
        "type": "object",
        "properties": {
          "manualResponsibleUnit": {
            "type": "string"
          },
          "force": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "InventorizationManualUnitResponseDto": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "updated"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsInventoryRequestDto": {
        "type": "object",
        "properties": {
          "inventoryOrderId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "accountSurrogateIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            },
            "description": "DOC: accountSurrogateKeys"
          },
          "manualInventoryAccountStatus": {
            "type": "string"
          },
          "discrepancyDescription": {
            "type": "string"
          },
          "discrepancySum": {
            "type": "string",
            "description": "Десятичное значение как строка (как в DOC string($decimal))"
          },
          "discrepancyReason": {
            "type": "string"
          },
          "resolutionActions": {
            "type": "string"
          },
          "resolutionDate": {
            "$ref": "#/components/schemas/DateSchema"
          }
        },
        "required": [
          "inventoryOrderId",
          "accountSurrogateIds"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsInventoryExcludeRequestDto": {
        "type": "object",
        "properties": {
          "inventoryOrderId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "accountSurrogateIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          }
        },
        "required": [
          "inventoryOrderId",
          "accountSurrogateIds"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsInventoryMutationResponseDto": {
        "type": "object",
        "properties": {
          "affected": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "affected"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountColumnDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "isVisible": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountColumnsResponseDto": {
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventorizationAccountColumnDto"
            }
          }
        },
        "required": [
          "columns"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountColumnsUpdateDto": {
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventorizationAccountColumnDto"
            }
          }
        },
        "required": [
          "columns"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsExportRequestDto": {
        "type": "object",
        "properties": {
          "accountSurrogateIds": {
            "description": "DOC: accountSurrogateKeys",
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "description": "Integer ID"
            }
          },
          "filter": {
            "$ref": "#/components/schemas/ШnventoryAccountsListFilterDto"
          },
          "inventoryOrderId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer ID"
          },
          "format": {
            "type": "string",
            "enum": [
              "xlsx",
              "csv"
            ]
          }
        },
        "required": [
          "filter"
        ],
        "additionalProperties": false
      },
      "InventorizationAccountsExportResponseDto": {
        "type": "object",
        "properties": {
          "fileKey": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/v1/inventorization/accounts/list": {
      "post": {
        "summary": "Список счетов инвентаризации (пагинация, сортировка, фильтры)",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationGetAccountsListRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountsListResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/columns": {
      "get": {
        "summary": "Колонки таблицы счетов",
        "tags": [
          "Inventorization - Accounts"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountColumnsResponseDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Сохранить настройки колонок",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationAccountColumnsUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountColumnsResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/export": {
      "post": {
        "summary": "Экспорт реестра счетов",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationAccountsExportRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountsExportResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/inventory": {
      "post": {
        "summary": "Включить счета в инвентаризацию",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationAccountsInventoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountsInventoryMutationResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/inventory/exclude": {
      "post": {
        "summary": "Исключить счета из инвентаризации",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationAccountsInventoryExcludeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountsInventoryMutationResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/manual-unit/bulk": {
      "post": {
        "summary": "Ручной учёт: массовая операция",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationManualUnitBulkRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationManualUnitResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/manual-unit/{accountSurrogateId}": {
      "post": {
        "summary": "Ручной учёт: одна запись (accountSurrogateId в URL)",
        "tags": [
          "Inventorization - Accounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorizationManualUnitSingleRequestDto"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "accountSurrogateId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationManualUnitResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/surrogate/{accountSurrogateId}/history": {
      "get": {
        "summary": "История изменений по surrogate id счёта",
        "tags": [
          "Inventorization - Accounts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "accountSurrogateId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountHistoryResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventorization/accounts/{id}": {
      "get": {
        "summary": "Деталь счёта по id",
        "tags": [
          "Inventorization - Accounts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorizationAccountDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api-stage.example.com",
      "description": "Staging server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
}